Unbreak the table
The proposed position="replace" is not a good practice and it is breaking my table when there is inheritance. Therefore I propose to add a new <t> tag and use position="move" to move the whole <tr> tag in <tbody> to the new tag. This is good for any further inheritance.